# services/api/apidocs/openapi.yaml
...
openapi: 3.1.0
info:
  title: SchoolERP API
  version: 1.0.0
  description: |
    Comprehensive API documentation for the SchoolERP platform.
    This API handles everything from student management, finance, and exams
    to AI-powered parent helpdesks and cross-campus analytics.

    ### Authentication
    Most endpoints require a signed JWT token in the `Authorization` header:
    `Authorization: Bearer <your_token_here>`

    ### Multi-tenancy
    The API supports multi-tenancy. You must specify the tenant ID in the
    `X-Tenant-ID` header or ensure the request is routed via the tenant's subdomain.

    ### Response Formats
    All responses are in JSON format and follow a consistent structure:
    - Success: `{ "success": true, "data": { ... } }`
    - Error: `{ "success": false, "message": "error description", "code": "ERR_CODE" }`
servers:
  - url: http://localhost:8080/v1
    description: Local development server
  - url: https://schoolerp-vmn1.onrender.com/v1
    description: Production API
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    TenantId:
      type: apiKey
      in: header
      name: X-Tenant-ID
security:
  - BearerAuth: []
    TenantId: []
paths:
...
  /admin/billing/credits/ledger:
    get:
      operationId: getCreditLedger
      tags: [Billing Credits]
      summary: Get tenant credit ledger entries
      parameters:
        - in: query
          name: wallet_type
          schema: { type: string, example: sms_credits }
        - in: query
          name: limit
          schema: { type: integer, default: 50 }
        - in: query
          name: offset
          schema: { type: integer, default: 0 }
      responses:
        '200':
          description: Credit ledger entries
  /admin/billing/credits/topup:
    post:
      operationId: requestCreditTopup
      tags: [Billing Credits]
      summary: Request a tenant credit top-up (platform approval/invoice flow)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [wallet_type, amount]
              properties:
                wallet_type: { type: string, example: sms_credits }
                amount: { type: integer, format: int64, example: 1000 }
      responses:
        '202':
          description: Top-up request created and pending approval
  /admin/platform/billing/credits/adjust:
    post:
      operationId: platformAdjustCredits
      tags: [Platform Billing]
      summary: Adjust tenant credits (support/finance action)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, wallet_type, amount, reference_id]
              properties:
                tenant_id: { type: string, format: uuid }
                wallet_type: { type: string }
                amount: { type: integer, format: int64, description: "Positive for credit, negative for debit" }
                source: { type: string, example: support_adjustment }
                reference_id: { type: string }
                metadata: { type: object }
      responses:
        '200':
          description: Adjustment applied
        '409':
          description: Insufficient credits for debit adjustment
...
  /admin/settings/integrations/{provider}/connect:
    post:
      operationId: startTenantIntegrationConnect
      tags: [Integrations]
      summary: Start OAuth connection for tenant integration
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google_workspace, microsoft_365] }
      responses:
        '200':
          description: OAuth start response with auth URL
  /admin/settings/integrations/{provider}/callback:
    get:
      operationId: finishTenantIntegrationConnectAdminCallback
      tags: [Integrations]
      summary: Admin-auth callback alias for OAuth completion
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google_workspace, microsoft_365] }
        - in: query
          name: state
          required: true
          schema: { type: string }
        - in: query
          name: code
          required: true
          schema: { type: string }
      responses:
        '302':
          description: Redirect to integrations page
  /integrations/oauth/{provider}/callback:
    get:
      operationId: finishTenantIntegrationConnectPublicCallback
      tags: [Integrations]
      summary: Public OAuth callback (tenant resolved by stored oauth_state)
      security: []
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google_workspace, microsoft_365] }
        - in: query
          name: state
          required: true
          schema: { type: string }
        - in: query
          name: code
          required: true
          schema: { type: string }
        - in: query
          name: redirect
          required: false
          schema: { type: string }
      responses:
        '302':
          description: Redirect to integrations page
  /admin/settings/integrations/{provider}/disconnect:
    post:
      operationId: disconnectTenantIntegration
      tags: [Integrations]
      summary: Disconnect tenant integration and clear stored tokens
      parameters:
        - in: path
          name: provider
          required: true
          schema: { type: string, enum: [google_workspace, microsoft_365] }
      responses:
        '200':
          description: Integration disconnected
...